Added import statement for Apple's UIKit framework as it is required dependency.
The Countly class uses methods from the UIKit framework however it doesn't import that framework. This wont be a problem for most apps because the default project template for Xcode imports the UIKit framework in the apps Prefix.pch. However when an application doesn't use the provided Prefix.pch the Countly class will fail to compile.