Allow usage of suspension mechanism for broader use.
It is useful to allow an application to request temporary manual suspension of event recording.

Currently, the only available mechanism to shut down Countly's tracking
once started is the consent mechanism. This mechanism is too
heavy-handed for temporary/situational suspension since it affects
global consent counters for the userbase.

The existing suspension mechanism is a great fit for temporarily
shutting down Countly and allowing Countly to resume later on.

Suggested use case:
    - A user enters incognito mode and the application needs to be able
      to make privacy guarantees.
    - The app wants to go offline and ensure that it is not emitting or
      triggering any network activity.